requirements.txt
pytesseract==0.3.10
pillow==9.5.0